week07/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
Bottleneck @ file:///croot/bottleneck_1731058641041/work
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.3.0
click==8.3.0
coloredlogs==15.0.1
contourpy @ file:///croot/contourpy_1732540045555/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools @ file:///home/task_175975467226191/conda-bld/fonttools_1759754871865/work
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.18
google-api-core==2.27.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.1.8
langchain-tavily==0.2.12
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.38
markdown-it-py==4.0.0
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///croot/mkl_fft_1751317925336/work
mkl_random @ file:///croot/mkl_random_1751318527880/work
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
numexpr @ file:///croot/numexpr_1755766469354/work
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas @ file:///home/task_175982152705186/conda-bld/pandas_1759822246405/work/dist/pandas-2.3.3-cp312-cp312-linux_x86_64.whl#sha256=594352c90592c883f6b80a8c3ab4eb2438ac1feda05bb5540854c0f37cb31267
pillow @ file:///croot/pillow_1752524526621/work
playwright==1.55.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
pyparsing @ file:///croot/pyparsing_1731445506121/work
pypdf==6.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt6==6.9.1
PyQt6_sip @ file:///home/task_175751008914507/conda-bld/pyqt-split_1757511022708/work/pyqt_sip
PySocks==1.7.1
pysqlite3-binary==0.5.4
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.1
pytz @ file:///croot/pytz_1752135852232/work
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
selenium==4.38.0
setuptools==80.9.0
shellingham==1.5.4
sip @ file:///croot/sip_1756223371714/work
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.12.0
tokenizers==0.22.1
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata @ file:///croot/python-tzdata_1746123641790/work
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
